@@ -13,18 +13,8 @@ use ThreeBRS\SyliusPaymentFeePlugin\Model\Calculator\DelegatingCalculatorInterfa
 
 final class PaymentChargesProcessor implements OrderProcessorInterface
 {
-    /** @var FactoryInterface */
-    private $adjustmentFactory;
-
-    /** @var DelegatingCalculatorInterface */
-    private $paymentChargesCalculator;
-
-    public function __construct(
-        FactoryInterface $adjustmentFactory,
-        DelegatingCalculatorInterface $paymentChargesCalculator
-    ) {
-        $this->adjustmentFactory = $adjustmentFactory;
-        $this->paymentChargesCalculator = $paymentChargesCalculator;
+    public function __construct(private FactoryInterface $adjustmentFactory, private DelegatingCalculatorInterface $paymentChargesCalculator)
+    {
     }
 
     public function process(BaseOrderInterface $order): void
